Create base class for HTTP client
